blueprints/examples/EKS/dags/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.2.2/constraints-3.7.txt"
apache-airflow-providers-amazon==3.0.0
apache-airflow-providers-cncf-kubernetes==3.0.0

